Get rid of most warning

+++ decomplz4.h
@@ -10,7 +10,7 @@
 // read one byte from input, see getByteFromIn()  for a basic implementation
 typedef unsigned char (*DECOMP_GET_BYTE)  (void* userPtr);
 // write several bytes,      see sendBytesToOut() for a basic implementation
-typedef void          (*DECOMP_SEND_BYTES)(const unsigned char*, unsigned int, void* userPtr);
+typedef void          (*DECOMP_SEND_BYTES)(const unsigned char*, uint64_t, void* userPtr);
 
 void unlz4_userPtr(DECOMP_GET_BYTE getByte, DECOMP_SEND_BYTES sendBytes, const char* dictionary, void* userPtr);
 void unlz4error(const char* msg);
